requirements.txt
mkdocs~=1.3.1
pyserial~=3.5
pytest~=6.2.4
pyusb~=1.2.1
streamexpect~=0.2.1
west~=0.13.1
